Inline useDocumentGateway
