+++ web/packages/teleterm/src/ui/DocumentGatewayApp/DocumentGatewayApp.tsx
@@ -15,12 +15,10 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
-import { useAppContext } from 'teleterm/ui/appContextProvider';
 import { DocumentGateway } from 'teleterm/ui/services/workspacesService';
 import Document from 'teleterm/ui/Document';
 
-import { useDocumentGateway } from '../DocumentGateway/useDocumentGateway';
+import { useGateway } from '../DocumentGateway/useGateway';
 
 import { OfflineGateway } from '../components/OfflineGateway';
 
@@ -30,7 +28,7 @@ export function DocumentGatewayApp(props: {
   doc: DocumentGateway;
   visible: boolean;
 }) {
-  const ctx = useAppContext();
+  const { doc } = props;
   const {
     gateway,
     changePort,
@@ -40,9 +38,7 @@ export function DocumentGatewayApp(props: {
     disconnect,
     disconnectAttempt,
     reconnect,
-  } = useDocumentGateway(props.doc);
-
-  ctx.clustersService.useState();
+  } = useGateway(doc);
 
   return (
     <Document visible={props.visible}>
@@ -50,8 +46,8 @@ export function DocumentGatewayApp(props: {
         <OfflineGateway
           connectAttempt={connectAttempt}
           gatewayKind="app"
-          targetName={props.doc.targetName}
-          gatewayPort={{ isSupported: true, defaultPort: props.doc.port }}
+          targetName={doc.targetName}
+          gatewayPort={{ isSupported: true, defaultPort: doc.port }}
           reconnect={reconnect}
         />
       ) : (
